Move DB interactions to REST API's (#5)
* Move logic to APIs

* List subscriptions for areas

* Move from geo area to target area selection

* Ensure subscription editing works with contact-api

Note: files and routes still refer to locations, this needs to change

* Fix delete subscription functionality

Also, switched to using postgrest rather than using a custom
contact-api. Because it's all just rest this should work with
any api which treats contacts and subscriptions as resouces.

Decided to have a distinct URL config value for each REST method to
allow Command Query Responsibility Separation (CQRS). This means that
if, for example, we want to raise an event everytime a new contact is
created we can use a small passthrough API app which will raise the
event as well as passing the call to postgrest

* Remove leaking of postgres naming into js code

Also:
* fixed issue contact page not showing added contact due to async/await
  error
* change code to reflect the area name are returned as an embedded resource
  of subscription

* Distinguish between FAA's and FWA's when selecting areas

* Fix null reference issue

* Move interpolate function to shared utils

* Area API now expects a coord _or_ bbox parameter

* Fix wreck version